+++ examples/evo_trader/run.py
@@ -2,10 +2,8 @@ from vnpy_evo.event import EventEngine
 from vnpy_evo.trader.engine import MainEngine
 from vnpy_evo.trader.ui import MainWindow, create_qapp
 
-from vnpy_binance import BinanceUsdtGateway
-from vnpy_ctastrategy import CtaStrategyApp
-from vnpy_ctabacktester import CtaBacktesterApp
-from vnpy_datamanager import DataManagerApp
+from vnpy_binance import BinanceLinearGateway
+from vnpy_novastrategy import NovaStrategyApp
 
 
 def main():
@@ -16,11 +14,9 @@ def main():
 
     main_engine = MainEngine(event_engine)
 
-    main_engine.add_gateway(BinanceUsdtGateway)
+    main_engine.add_gateway(BinanceLinearGateway)
 
-    main_engine.add_app(CtaStrategyApp)
-    main_engine.add_app(CtaBacktesterApp)
-    main_engine.add_app(DataManagerApp)
+    main_engine.add_app(NovaStrategyApp)
 
     main_window = MainWindow(main_engine, event_engine)
     main_window.showMaximized()