+++ install_macos.sh
@@ -1,0 +1,14 @@
+#!/bin/zsh
+
+# Install NumPy
+pip3 install numpy==1.26.4
+
+# Install ta-lib
+brew install ta-lib
+pip3 install ta-lib
+
+# Install vnpy_evo
+pip3 install vnpy_evo
+pip3 install vnpy_duckdb
+pip3 install vnpy_binance vnpy_btse vnpy_okx vnpy_bybit
+pip3 install vnpy_novastrategy