+++ CHANGELOG.md
@@ -1,3 +1,37 @@
+# 0.3.1
+
+1. add new vnpy_evo.rest as alternative of vnpy_rest
+2. add new vnpy_evo.websocket as alternative of vnpy_websocket
+
+# 0.3.0
+
+1. New vnpy_btse gateway module
+2. New vnpy_novastrategy module for crypto market quant strategies
+3. New vnpy_duckdb database module
+4. Quick install script for Windows/Linux/Mac OS
+
+# 0.2.2
+
+1. reimplement extract_vt_symbol function
+
+# 0.2.1
+
+1. add BTSE exchange num
+
+# 0.2.0
+
+## Gateway
+
+1. Update vnpy_mt5 gateway module
+2. Update vnpy_okx gateway module
+3. Update vnpy_bybit gateway module
+
+## Framwork
+
+1. Add new OTC exchange enum
+2. Only support English language for i18n
+3. Fix bugs related to qfluentwidgets
+
 # 0.1.0
 
 1. The first release of VeighNa Evo